+++ source/lib/constants.ts
@@ -1,0 +1,46 @@
+/*********************************************************************************************************************
+ *  Copyright 2020-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                      *
+ *                                                                                                                    *
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
+ *  with the License. A copy of the License is located at                                                             *
+ *                                                                                                                    *
+ *      http://www.apache.org/licenses/LICENSE-2.0                                                                    *
+ *                                                                                                                    *
+ *  or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES *
+ *  OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions    *
+ *  and limitations under the License.                                                                                *
+ *********************************************************************************************************************/
+
+export default class Constants {
+    private static readonly ssmManagedActions = [
+        'ssm:DescribeAssociation',
+        'ssm:GetDeployablePatchSnapshotForInstance',
+        'ssm:GetDocument',
+        'ssm:DescribeDocument',
+        'ssm:GetManifest',
+        'ssm:GetParameter',
+        'ssm:GetParameters',
+        'ssm:ListAssociations',
+        'ssm:ListInstanceAssociations',
+        'ssm:PutInventory',
+        'ssm:PutComplianceItems',
+        'ssm:PutConfigurePackageResult',
+        'ssm:UpdateAssociationStatus',
+        'ssm:UpdateInstanceAssociationStatus',
+        'ssm:UpdateInstanceInformation',
+        'ec2messages:AcknowledgeMessage',
+        'ec2messages:DeleteMessage',
+        'ec2messages:FailMessage',
+        'ec2messages:GetEndpoint',
+        'ec2messages:GetMessages',
+        'ec2messages:SendReply',
+        'ssmmessages:CreateControlChannel',
+        'ssmmessages:CreateDataChannel',
+        'ssmmessages:OpenControlChannel',
+        'ssmmessages:OpenDataChannel'
+    ]
+
+    public static get SsmManagedActions(): string[] {
+        return Constants.ssmManagedActions;
+    }
+}