+++ source/lib/msk-client.ts
@@ -16,6 +16,7 @@ import { Construct } from 'constructs';
 import { aws_ec2 as ec2, aws_iam as iam } from 'aws-cdk-lib';
 
 import { CfnNagHelper } from './cfn-nag-helper';
+import Constants from './constants';
 
 export interface KafkaClientProps {
     readonly vpcId: string;
@@ -109,16 +110,10 @@ export class KafkaClient extends Construct {
             assumedBy: new iam.ServicePrincipal('ec2.amazonaws.com')
         });
 
-        const ssmPolicy = new iam.Policy(this, 'SessionManagerPolicy', {
+        const ssmPolicy = new iam.Policy(this, 'SessionManagerPolicy', { // NOSONAR: Explicitly listing actions is security standard
             statements: [new iam.PolicyStatement({
                 resources: ['*'],
-                actions: [
-                    'ssm:UpdateInstanceInformation',
-                    'ssmmessages:CreateControlChannel',
-                    'ssmmessages:CreateDataChannel',
-                    'ssmmessages:OpenControlChannel',
-                    'ssmmessages:OpenDataChannel'
-                ]
+                actions: Constants.SsmManagedActions
             })]
         });