+++ source/lib/msk-cluster.ts
@@ -46,6 +46,12 @@ export enum KafkaInstanceType {
 }
 
 export enum KafkaActiveVersion {
+    V3_5_1 = '3.5.1',
+    V3_4_0 = '3.4.0',
+    V3_3_2 = '3.3.2',
+    V3_3_1 = '3.3.1',
+    V3_2_0 = '3.2.0',
+    V3_1_1 = '3.1.1',
     V2_8_1 = '2.8.1',
     V2_8_0 = '2.8.0',
     V2_7_2 = '2.7.2',
@@ -56,7 +62,7 @@ export enum KafkaActiveVersion {
     V2_6_1 = '2.6.1',
     V2_6_0 = '2.6.0',
     V2_5_1 = '2.5.1',
-    V2_4_1_1 = '2.4.1.1',
+    V2_4_1_1 = '2.4.1.1', // NOSONAR: not an IP address
     V2_3_1 = '2.3.1',
     V2_2_1 = '2.2.1'
 }